server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
together>=1.0.0
twilio==8.10.0
python-dateutil==2.8.2
requests==2.31.0
textblob
nltk